feat(GiniHealthSDK): Refactor Payment Review Container info into a programatically view
