+++ MerchantSDK/GiniMerchantSDK/Sources/GiniMerchantSDK/Core/PaymentReviewViewController+PaymentReviewViewModelDelegate.swift
@@ -14,7 +14,7 @@ extension PaymentReviewViewController: PaymentReviewViewModelDelegate {
 
     func createPaymentRequestAndOpenBankApp() {
         self.presentedViewController?.dismiss(animated: true)
-        checkForErrors()
+        paymentInfoContainerView.noErrorsFound()
     }
 
     func presentShareInvoiceBottomSheet(bottomSheet: BottomSheetViewController) {
@@ -24,7 +24,7 @@ extension PaymentReviewViewController: PaymentReviewViewModelDelegate {
     }
 
     func obtainPDFFromPaymentRequest() {
-        model?.createPaymentRequest(paymentInfo: obtainPaymentInfo(), completion: { [weak self] paymentRequestID in
+        model?.createPaymentRequest(paymentInfo: paymentInfoContainerView.obtainPaymentInfo(), completion: { [weak self] paymentRequestID in
             self?.loadPDFData(paymentRequestID: paymentRequestID)
         })
     }
